Simple cache stats simulator
I tried a few different approaches before settling on using awk for the
main processing. Ruby was too slow. Also tried loading the data into
sqlite. Loading is fast but it's hard to do the right manipulation in
sqlite.

Once I realised that the primary processing of the data is chronological
by url, it seemed obvious that feeding sorted logfiles to awk would be
the best strategy.